refactor: CalendarFragment.kt -> HomeFragment.kt로 이동
