+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.GagyeBoost" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.GagyeBoost" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
@@ -19,6 +19,7 @@
         <item name="android:layout_width">0dp</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="android:textSize">10sp</item>
+        <item name="android:layout_weight">1</item>
         <item name="android:layout_marginTop">8dp</item>
         <item name="android:gravity">center</item>
     </style>
